+++ examples/using-cypress/gatsby-ssr.js
@@ -1,3 +1,0 @@
-exports.onRenderBody = ({ setHtmlAttributes }) => {
-  setHtmlAttributes({ lang: `en` })
-}