+++ packages/gatsby-cypress/src/commands.js
@@ -2,6 +2,7 @@
 
 import apiHandler, { waitForAPI } from "./api-handler"
 
+// TODO(v6): Remove this command
 Cypress.Commands.add(`getTestElement`, (selector, options = {}) =>
   cy.get(`[data-testid="${selector}"]`, options)
 )