+++ packages/gatsby-source-contentful/src/fetch.ts
@@ -332,7 +332,7 @@ export async function fetchContent({
           resolveLinks: false,
         }
         const query = syncToken
-          ? { nextSyncToken: syncToken, ...basicSyncConfig }
+          ? { nextSyncToken: syncToken }
           : { initial: true, ...basicSyncConfig }
         currentSyncData = (await syncClient.sync(query)) as SyncCollection
         syncSuccess = true
@@ -378,8 +378,7 @@ export async function fetchContent({
       currentSyncData?.entries.length +
         currentSyncData?.assets.length +
         currentSyncData?.deletedEntries.length +
-        currentSyncData?.deletedAssets.length ===
-        0
+        currentSyncData?.deletedAssets.length
     ) {
       syncProgress.tick()
       syncProgress.total = 1