+++ packages/gatsby-source-graphql/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.10.0](https://github.com/gatsbyjs/gatsby/commits/gatsby-source-graphql@5.10.0/packages/gatsby-source-graphql) (2023-05-16)
+
+[🧾 Release notes](https://www.gatsbyjs.com/docs/reference/release-notes/v5.10)
+
+#### Bug Fixes
+
+- update minor and patch dependencies for gatsby-source-graphql [#38028](https://github.com/gatsbyjs/gatsby/issues/38028) ([d611439](https://github.com/gatsbyjs/gatsby/commit/d61143994682727693eb91e971bfa1634cee1326))
+
 ## [5.9.0](https://github.com/gatsbyjs/gatsby/commits/gatsby-source-graphql@5.9.0/packages/gatsby-source-graphql) (2023-04-18)
 
 [🧾 Release notes](https://www.gatsbyjs.com/docs/reference/release-notes/v5.9)